% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_aoa.R
\name{calibrate_aoa}
\alias{calibrate_aoa}
\title{Calibrate the AOA based on the relationship between the DI and the prediction error}
\usage{
calibrate_aoa(AOA, model, calibration = NULL, response = NULL, showplot = TRUE)
}
\arguments{
\item{AOA}{the result of ?aoa}

\item{model}{the model used to get the AOA}

\item{calibration}{Optional. Data frame containing a new set of predictor and response values.}

\item{response}{character. Name of the response variable used in the model and present in the calibration data frame. Only required when calibration data are given.}

\item{showplot}{Logical. Show visualization oder not?}
}
\value{
rasterStack which contains the original DI and the AOA (whoch might be updated if new test data indicate this option), as well as the expected error.
}
\description{
Calibrate the AOA based on the relationship between the DI and the prediction error
}
\details{
The AOA is the area to which the model can be applied and where the reliability of predictions is expected to be comparable to the cross-validation error of the model.
However, it might be desireable to limit predictions to an area with a user-defined error.
This function allows for this based on the relationship of the DI an the absolute prediction error derived from cross-validation during model training.

The error-DI relationship can also be estimated by providing new external calibration data points.
In this case the DI is calculated from these data and the error~DI relationship is derived.
If the AOA threshold based on the calibration data is larger than the original AOA threshold, the AOA is updated accordingly.
}
\examples{
\dontrun{
library(sf)
library(raster)
library(caret)
library(viridis)
library(latticeExtra)

# prepare sample data:
library(sf)
library(raster)
library(caret)
# prepare sample data:
dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
dat <- aggregate(dat[,c("VW","Easting","Northing")],by=list(as.character(dat$SOURCEID)),mean)
pts <- st_as_sf(dat,coords=c("Easting","Northing"))
pts$ID <- 1:nrow(pts)
studyArea <- stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))[[1:8]]
dat <- extract(studyArea,pts,df=TRUE)
trainDat <- merge(dat,pts,by.x="ID",by.y="ID")

# train a model:
variables <- c("DEM","NDRE.Sd","TWI")
set.seed(100)
model <- train(trainDat[,which(names(trainDat)\%in\%variables)],
  trainDat$VW,method="rf",importance=TRUE,tuneLength=1,
  trControl=trainControl(method="cv",number=5,savePredictions=TRUE))

#...then calculate the AOA of the trained model for the study area:
AOA <- aoa(studyArea,model)

AOA_new <- calibrate_aoa(AOA,model)
plot(AOA_new$expectedError)


####example with new calibration data:
#...to do

}
}
\seealso{
\code{\link{aoa}}
}
\author{
Hanna Meyer
}
