% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAST-package.R
\docType{package}
\name{CAST}
\alias{CAST}
\alias{CAST-package}
\title{Caret Applications for Spatio-Temporal models}
\description{
The package includes functionality to run caret with spatial or spatio-temporal data. The package includes the functions published in \href{https://doi.org/10.1016/j.envsoft.2017.12.001}{Meyer et al. (2018)}.
}
\details{
Caret Applications for Spatio-Temporal models
}
\author{
Hanna Meyer
\cr
\emph{Maintainer:} Hanna Meyer \email{hanna.meyer@geo.uni-marburg.de}
}
\keyword{package}
