% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DItoErrormetric.R
\name{errorModel_DI}
\alias{errorModel_DI}
\title{Model expected error between Metric and DI}
\usage{
errorModel_DI(preds_all, model, window.size, calib, k, m)
}
\arguments{
\item{preds_all}{data.frame: pred, obs, DI}

\item{model}{the model used to get the AOA}

\item{window.size}{Numeric. Size of the moving window. See \code{\link{rollapply}}.}

\item{calib}{Character. Function to model the DI~performance relationship. Currently lm and scam are supported}

\item{k}{Numeric. See mgcv::s}

\item{m}{Numeric. See mgcv::s}
}
\value{
scam or lm
}
\description{
Model expected error between Metric and DI
}
