% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI_LPDtoErrormetric.R
\name{errorModel_DI_LPD}
\alias{errorModel_DI_LPD}
\title{Model expected error between Metric and DI and LPD}
\usage{
errorModel_DI_LPD(preds_all, model, window.size, calib, kDI, mDI, kLPD, mLPD)
}
\arguments{
\item{preds_all}{data.frame: pred, obs, DI, LPD}

\item{model}{the model used to get the AOA}

\item{window.size}{Numeric. Size of the moving window. See \code{\link{rollapply}}.}

\item{calib}{Character. Function to model the DI~performance relationship. Currently lm and scam are supported}

\item{k}{Numeric. See mgcv::s}

\item{m}{Numeric. See mgcv::s}
}
\value{
scam or lm
}
\description{
Model expected error between Metric and DI and LPD
}
