% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPDtoErrormetric.R
\name{errorModel_LPD}
\alias{errorModel_LPD}
\title{Model expected error between Metric and LPD}
\usage{
errorModel_LPD(preds_all, model, window.size, calib, k, m)
}
\arguments{
\item{preds_all}{data.frame: pred, obs, LPD}

\item{model}{the model used to get the AOA}

\item{window.size}{Numeric. Size of the moving window. See \code{\link{rollapply}}.}

\item{calib}{Character. Function to model the LPD~performance relationship. Currently lm, scam and nls are supported}

\item{k}{Numeric. See mgcv::s}

\item{m}{Numeric. See mgcv::s}
}
\value{
scam, lm or nls model
}
\description{
Model expected error between Metric and LPD
}
