% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreAOA.R
\name{exploreAOA}
\alias{exploreAOA}
\title{Explore Area of Applicability}
\usage{
exploreAOA(aoa)
}
\arguments{
\item{aoa}{An R object of class \code{aoa} (generated by the \code{\link{aoa}} function).}
}
\description{
This function lets the user explore an calculated Area of Applicability object
interactively in an R shiny application. Depending on whether the parameter \code{LPD = TRUE}
was specified in the calculation of the \code{AOA}, the user can either explore the
\code{AOA} and \code{DI} layer or additionally explore the \code{LPD} and a reactive \code{neighbor-dependent
AOA layer}.
The functionality is very suitable for a first visual interpretation of the
Area of Applicability.
}
\details{
Interpretation of results: If a location is very similar to the properties
of the training data it will have a low distance in the predictor variable space
(\code{DI} towards 0) while locations that are very different in their properties
will have a high \code{DI}.
Depending on a threshold for the DI, the location will either be classified as
\code{Inside AOA (1)} or \code{Outside AOA (0)}. The comparison is done for the nearest neighbor
of the training points in the predictor space. This can be seen in the \code{AOA} layer.
The \code{LPD} (Local Data Point Density) layer indicates the total number of training data points
for which the locations \code{DI} is smaller than the threshold, i.e. on how many similar training points the model was trained.
The fourth layer is based on the \code{LPD} layer and can be seen as a \code{neighbor-dependet AOA}. This layer extends the \code{AOA}
by a minimum number k of training data points for which the threshold condition must hold in order to be classified as \code{Indside AOA (1)}.
}
\examples{
\dontrun{

}
}
\references{
Meyer, H., Pebesma, E. (2021): Predicting into unknown space?
Estimating the area of applicability of spatial prediction models.
Methods in Ecology and Evolution 12: 1620-1633. \doi{10.1111/2041-210X.13650}
}
\author{
Fabian Schumacher
}
