% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DItoErrormetric.R
\name{get_preds_all}
\alias{get_preds_all}
\title{Get Preds all}
\usage{
get_preds_all(model, trainDI, what)
}
\arguments{
\item{model, }{a model}

\item{trainDI, }{a trainDI}

\item{what}{Character. Which measure to use for the error metric. "DI" or "LPD"}
}
\description{
Get Preds all
}
