% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DItoErrormetric.R
\name{get_preds_all_DI}
\alias{get_preds_all_DI}
\title{Get Preds all}
\usage{
get_preds_all_DI(model, trainDI)
}
\arguments{
\item{model, }{a model}

\item{trainDI, }{a trainDI}
}
\description{
Get Preds all
}
