% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPDtoErrormetric.R
\name{get_preds_all_LPD}
\alias{get_preds_all_LPD}
\title{Get Preds all}
\usage{
get_preds_all_LPD(model, trainDI)
}
\arguments{
\item{model, }{a model}

\item{trainDI, }{a trainDI}
}
\description{
Get Preds all
}
