% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI_LPDtoErrormetric.R
\name{multiCV_DI_LPD}
\alias{multiCV_DI_LPD}
\title{MultiCV}
\usage{
multiCV_DI_LPD(model, length.out, method, useWeight, ...)
}
\arguments{
\item{model}{the model used to get the AOA}

\item{length.out}{Numeric. Only used if multiCV=TRUE. Number of cross-validation folds. See details.}

\item{method}{Character. Method used for distance calculation. Currently euclidean distance (L2) and Mahalanobis distance (MD) are implemented but only L2 is tested. Note that MD takes considerably longer. See ?aoa for further explanation}

\item{useWeight}{Logical. Only if a model is given. Weight variables according to importance in the model?}

\item{...}{additional parameters to trainDI}
}
\value{
preds_all
}
\description{
Multiple Cross-Validation with increasing feature space clusteres
}
