% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.aoa}
\alias{plot.aoa}
\title{Plot AOA}
\usage{
\method{plot}{aoa}(x, samplesize = 1000, ...)
}
\arguments{
\item{x}{aoa object}

\item{samplesize}{numeric. How many prediction samples should be plotted?}

\item{...}{other params}
}
\description{
Density plot of the DI with AOA threshold
}
\author{
Marvin Ludwig, Hanna Meyer
}
