% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.trainDI}
\alias{plot.aoa}
\title{Plot CAST classes}
\usage{
\method{plot}{trainDI}(x, ...)

\method{plot}{aoa}(x, samplesize = 1000, ...)
}
\arguments{
\item{x}{aoa object}

\item{...}{other params}

\item{samplesize}{numeric. How many prediction samples should be plotted?}
}
\description{
Generic plot function for trainDI and aoa
}
\author{
Marvin Ludwig, Hanna Meyer

Marvin Ludwig, Hanna Meyer
}
