% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trainDI}
\alias{plot.trainDI}
\title{Plot TrainDI}
\usage{
\method{plot}{trainDI}(x, ...)
}
\arguments{
\item{x}{trainDI object}

\item{...}{other params}
}
\description{
Plot TrainDI
}
\author{
Marvin Ludwig, Hanna Meyer
}
