% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ffs.R
\name{plot_ffs}
\alias{plot_ffs}
\title{Plot results of a Forward feature selection}
\usage{
plot_ffs(ffs_model, palette = "viridis")
}
\arguments{
\item{ffs_model}{Result of a forward feature selection see \code{\link{ffs}}}

\item{palette}{A color palette}
}
\description{
A simple plotting function for a forward feature selection result.
bars represent standard errors from cross validation.
Marked points show the best model from each number of variables.
}
\examples{
\dontrun{
data(iris)
ffsmodel <- ffs(iris[,1:4],iris$Species)
plot_ffs(ffsmodel)
}
}
\seealso{
\code{\link{ffs}}
}
\author{
Marvin Ludwig and Hanna Meyer
}
