% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.aoa}
\alias{print.aoa}
\alias{show.aoa}
\title{Print AoA objects}
\usage{
\method{print}{aoa}(x, ...)

show.aoa(x, ...)
}
\arguments{
\item{x}{aoa object}

\item{...}{other params}
}
\description{
Generic print function for objects of class \code{aoa}.
}
