% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ffs}
\alias{print.ffs}
\alias{show.ffs}
\title{Print ffs objects}
\usage{
\method{print}{ffs}(x, ...)

show.ffs(x, ...)
}
\arguments{
\item{x}{An object of type \emph{ffs}}

\item{...}{other arguments.}
}
\description{
Generic print function for objects of class \code{ffs}.
}
