% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.knndm}
\alias{print.knndm}
\alias{show.knndm}
\title{Print knndm objects}
\usage{
\method{print}{knndm}(x, ...)

show.knndm(x, ...)
}
\arguments{
\item{x}{An object of type \emph{knndm}.}

\item{...}{other arguments.}
}
\description{
Generic print function for objects of class \code{knndm}.
}
