% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nndm}
\alias{print.nndm}
\alias{show.nndm}
\title{Print nndm objects}
\usage{
\method{print}{nndm}(x, ...)

show.nndm(x, ...)
}
\arguments{
\item{x}{An object of type \emph{nndm}.}

\item{...}{other arguments.}
}
\description{
Generic print function for objects of class \code{nndm}.
}
