% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.trainDI}
\alias{show.trainDI}
\alias{print.aoa}
\alias{show.aoa}
\alias{print.nndm}
\alias{show.nndm}
\title{Print CAST classes}
\usage{
\method{print}{trainDI}(x, ...)

show.trainDI(x, ...)

\method{print}{aoa}(x, ...)

show.aoa(x, ...)

\method{print}{nndm}(x, ...)

show.nndm(x, ...)
}
\arguments{
\item{x}{An object of type \emph{nndm}.}

\item{...}{other arguments.}
}
\description{
Generic print function for trainDI and aoa
}
