% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.trainDI}
\alias{print.trainDI}
\alias{show.trainDI}
\title{Print trainDI objects}
\usage{
\method{print}{trainDI}(x, ...)

show.trainDI(x, ...)
}
\arguments{
\item{x}{trainDI object}

\item{...}{other params}
}
\description{
Generic print function for objects of class \code{trainDI}.
}
