% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc}
\alias{rc}
\alias{rc_vec}
\alias{rc.data.frame}
\alias{rc.train}
\alias{rc.aoa}
\title{Calculate the risk-coverage curve of a selective prediction model}
\usage{
rc_vec(score, residuals, risk = c("generalized", "selective"), n_bins = 100L)

rc(x, ...)

\method{rc}{data.frame}(x, risk = c("generalized", "selective"), loss = NULL, n_bins = 100L, ...)

\method{rc}{train}(x, aoa, score, ...)

\method{rc}{aoa}(x, model, residuals, ...)
}
\arguments{
\item{score}{A numeric vector of confidence scores. Lower values represent higher confidence.}

\item{residuals}{A numeric vector of residuals. Lower values represent smaller errors.}

\item{risk}{A character indicating the type of risk to calculate. Must be one of
\code{'generalized'} or \code{'selective'}.}

\item{n_bins}{An integer indicating the number of coverage bins between \code{[0,1]}.}

\item{x}{A supported object for which an \code{rc} method is available. See
more information in the details section.}

\item{...}{Arguments passed to the data.frame method. There, it is forwarded to
the loss-function, if specified. Otherwise ignored.}

\item{loss}{By default, \code{NULL}. Can be set to a function. In this case,
it is expected that the function receives at least the arguments \code{pred}
and \code{obs} and returns a scalar. Additional arguments are forwarded using
\code{...}.}

\item{aoa}{An (optional) object of type \code{\link{aoa}}. If not specified,
\code{score} is required.}

\item{model}{An (optional) object of type \code{\link[caret]{train}}. If not
specified, \code{residuals} is required.}
}
\value{
A \code{\link{data.frame}} of class \code{rc} containing the columns
  \code{coverage}, \code{empirical}, \code{reference}, and \code{excess}. The
  estimates of the respective AUC values can be accessed via \code{attr(rc, "aurc")}.
}
\description{
The risk-coverage curve describes the behavior of a model that is combined with
a confidence score function used to select samples to exclude from prediction.
Compared to traditional reporting of risk at full coverage (1.0), the
risk-coverage curve describes the trade-off behavior of a selective prediction
model across different levels of coverage and their associated risks.

The Area under the Risk-Coverage Curve (AURC) describes this trade-off
as a single scalar value. To make different models more comparable, the
excess-AURC (E-AURC) describes the excess area of an empirical risk-coverage
curve compared to a reference curve based on an optimal score function.
More information are presented in the details section and can be found in the linked
reference.
}
\details{
Note, that if \code{x} is a \code{data.frame}, it is required to
contain the columns \code{'score'} and \code{'residuals'}. In case \code{loss}
is specified, instead of the \code{'residuals'} column, the expected columns
are \code{'pred'} and \code{'obs'}. These columns are automatically retrieved
in the methods for other classes, or can be supplied manually.

Given a fitted model \eqn{f} and a selection function \eqn{g}, the
combined selective prediction model is defined as:

\deqn{
  (f,g)(x) \equiv \begin{cases}
  \text{$f(x)$, if $g(x) = 1$;}\\
  \text{reject, if $g(x) = 0$.}
  \end{cases}
}

The empirical coverage of such a model over a given data set
\eqn{S_m = \{x_i,y_i\}_{i=1}^m} is defined as

\deqn{
  \hat{\phi}(f,g|S_m) = \frac{1}{m} \sum_{i=1}^{m}g(x_i).
}

While we leverage an arbitrary loss function \eqn{\mathcal{L}} to derive its
empirical selective risk

\deqn{
  \hat r_s(f|S_m) = \frac{\frac{1}{m} \sum_{i=1}^{m} \mathcal{L}(f(x_i), y_i)g(x_i)}{\hat{\phi}(f,g|S_m)},
}
which serves as an estimator for the conditional probability \eqn{P(f(x) \ne y|g(x)=1)},
e.g. the risk of an undetected silent failure.

The empirical generalized risk is defined as

\deqn{
 \hat r_g(f|S_m) = \frac{1}{m} \sum_{i=1}^{m} \mathcal{L}(f(x_i), y_i)g(x_i),
}
which estimates the joint probability of prediction failure and sample
acceptance \eqn{P(f(x) \ne y, g(x)=1)}.

The risk-coverage curve then describes the performance profile of such a selective
prediction model, defining risk as a function of coverage. The curve can be
derived for any confidence score function \eqn{\kappa} inducing a selection
function \eqn{g} by applying a threshold \eqn{\lambda}:

\deqn{
  g_{\lambda}(x|\kappa,f) = \mathbb{1}[\kappa(x|f)>\lambda].
}
}
\examples{
library(CAST)

set.seed(42)
obs <- runif(100)
pred <- obs + (runif(n = 100, max = 10) * obs)
residuals <- abs(pred - obs) # MAE
score <- runif(n = 100) * residuals # decreasing confidence with residuals

# vectorized version with pre-calculated residuals
rc1 <- rc_vec(score = score, residuals = residuals)
# data.frame with pre-calculated residuals
rc2 <- rc(data.frame(score = score, residuals = residuals))
# data.frame with a specified loss function
rc3 <- rc(data.frame(pred = pred, obs = obs, score = score), loss = caret::MAE)

identical(rc1, rc2) & identical(rc1, rc3)

print(rc1)
attr(rc1, "auc")
plot(rc1)
}
\references{
Geifman, Y., Uziel, G., El-Yaniv, R., 2019. Bias-Reduced Uncertainty
  Estimation for Deep Neural Classifiers. https://doi.org/10.48550/arXiv.1805.08206

  Traub, J., Bungert, T.J., Lüth, C.T., Baumgartner, M., Maier-Hein, K.H., Maier-Hein,
  L., Jaeger, P.F., 2024. Overcoming Common Flaws in the Evaluation of Selective
  Classification Systems. https://doi.org/10.48550/arXiv.2407.01032
}
\author{
Darius A. Görgen
}
