% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAOA.R
\name{userAOA}
\alias{userAOA}
\title{User defined Area of Applicability}
\usage{
userAOA(AOA, model, proficiency = NULL, showplot = TRUE)
}
\arguments{
\item{AOA}{the result of ?aoa}

\item{model}{the model used to get the AOA}

\item{proficiency}{numeric. error accepted for the AOA. To first get an overview leave it to NA and select it after inspecting the plot}

\item{showplot}{Logical. Show visualization oder not?}
}
\value{

}
\description{
User defined Area of Applicability
}
\examples{
\dontrun{
library(sf)
library(raster)
library(caret)
library(viridis)
library(latticeExtra)

# prepare sample data:
dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
dat <- aggregate(dat[,c("VW","Easting","Northing")],by=list(as.character(dat$SOURCEID)),mean)
pts <- st_as_sf(dat,coords=c("Easting","Northing"))
pts$ID <- 1:nrow(pts)
set.seed(100)
pts <- pts[1:30,]
studyArea <- stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))[[1:8]]
trainDat <- extract(studyArea,pts,df=TRUE)
trainDat <- merge(trainDat,pts,by.x="ID",by.y="ID")

# train a model:
set.seed(100)
variables <- c("DEM","NDRE.Sd","TWI")
model <- train(trainDat[,which(names(trainDat)\%in\%variables)],
trainDat$VW,method="rf",importance=TRUE,tuneLength=1,
trControl=trainControl(method="cv",number=5,savePredictions=TRUE))

#...then calculate the AOA of the trained model for the study area:
AOA <- aoa(studyArea,model)
#...or get the area for a user-defined performance
AOA_new <- userAOA(AOA,model,proficiency=0.1)
}
}
\author{
Hanna Meyer
}
