% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.trainDI}
\alias{plot.aoa}
\alias{plot.nndm}
\alias{plot.knndm}
\alias{plot.ffs}
\title{Plot CAST classes}
\usage{
\method{plot}{trainDI}(x, ...)

\method{plot}{aoa}(x, samplesize = 1000, ...)

\method{plot}{nndm}(x, ...)

\method{plot}{knndm}(x, ...)

\method{plot}{ffs}(
  x,
  plotType = "all",
  palette = rainbow,
  reverse = FALSE,
  marker = "black",
  size = 1.5,
  lwd = 0.5,
  pch = 21,
  ...
)
}
\arguments{
\item{x}{Result of a forward feature selection see \code{\link{ffs}}}

\item{...}{Further arguments for base plot if type="selected"}

\item{samplesize}{numeric. How many prediction samples should be plotted?}

\item{plotType}{character. Either "all" or "selected"}

\item{palette}{A color palette}

\item{reverse}{Character. Should the palette be reversed?}

\item{marker}{Character. Color to mark the best models}

\item{size}{Numeric. Size of the points}

\item{lwd}{Numeric. Width of the error bars}

\item{pch}{Numeric. Type of point marking the best models}
}
\description{
Generic plot function for trainDI and aoa

A plotting function for a forward feature selection result.
Each point is the mean performance of a model run. Error bars represent
the standard errors from cross validation.
Marked points show the best model from each number of variables until a further variable
could not improve the results.
If type=="selected", the contribution of the selected variables to the model
performance is shown.
}
\examples{
\dontrun{
data(iris)
ffsmodel <- ffs(iris[,1:4],iris$Species)
plot(ffsmodel)
#plot performance of selected variables only:
plot(ffsmodel,plotType="selected")
}
}
\seealso{
\code{\link{ffs}}, \code{\link{bss}}
}
\author{
Marvin Ludwig, Hanna Meyer

Carles Milà

Marvin Ludwig and Hanna Meyer
}
