% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.trainDI}
\alias{plot.aoa}
\alias{plot.nndm}
\alias{plot.knndm}
\title{Plot CAST classes}
\usage{
\method{plot}{trainDI}(x, ...)

\method{plot}{aoa}(x, samplesize = 1000, ...)

\method{plot}{nndm}(x, ...)

\method{plot}{knndm}(x, ...)
}
\arguments{
\item{x}{An object of type \emph{knndm}.}

\item{...}{other arguments.}

\item{samplesize}{numeric. How many prediction samples should be plotted?}
}
\description{
Generic plot function for trainDI and aoa
}
\author{
Marvin Ludwig, Hanna Meyer

Carles Milà
}
