% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.trainDI}
\alias{show.trainDI}
\alias{print.aoa}
\alias{show.aoa}
\title{Print}
\usage{
\method{print}{trainDI}(x, ...)

show.trainDI(x, ...)

\method{print}{aoa}(x, ...)

show.aoa(x, ...)
}
\arguments{
\item{x}{aoa object}

\item{...}{other params}
}
\description{
Generic print function for trainDI and aoa
}
