% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainDI.R
\name{trainDI}
\alias{trainDI}
\title{Dissimilarity Index}
\usage{
trainDI(
  model,
  train = NULL,
  variables = "all",
  weight = NA,
  folds = NULL,
  returnTrainDI = TRUE
)
}
\arguments{
\item{model}{A train object created with caret used to extract weights from (based on variable importance) as well as cross-validation folds}

\item{train}{A data.frame containing the data used for model training. Only required when no model is given}

\item{variables}{character vector of predictor variables. if "all" then all variables
of the model are used or if no model is given then of the train dataset.}

\item{weight}{A data.frame containing weights for each variable. Only required if no model is given.}

\item{folds}{Numeric or character. Folds for cross validation. E.g. Spatial cluster affiliation for each data point.
Should be used if replicates are present. Only required if no model is given.}

\item{returnTrainDI}{A logical: should the DI value of the cross-validated training data be returned as a attribute?}
}
\value{
A trainDI object
}
\description{
This function estimates the Dissimilarity Index (DI) of
of spatial prediction models. Predictors can be weighted based on the internal
variable importance of the machine learning algorithm used for model training.
}
\references{
Meyer, H., Pebesma, E. (2020): Predicting into unknown space?
Estimating the area of applicability of spatial prediction models.
\url{https://arxiv.org/abs/2005.07939}
}
\seealso{
\code{\link{aoa}}
}
\author{
Hanna Meyer, Marvin Ludwig
}
