% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{uncertainty}
\alias{uncertainty}
\title{Estimate uncertainty of spatial prediction models}
\usage{
uncertainty(train, predictors, weights = NA, model = NA,
  variables = "all")
}
\arguments{
\item{train}{a data.frame containing the data used for model training}

\item{predictors}{A RasterStack, RasterBrick or data.frame containing the data
the model was meant to make predictions for.}

\item{weights}{A data.frame containing weights for each variable}

\item{model}{A caret model used to extract weights from (based on variable importance)}

\item{variables}{character vector of predictor variables. if "all" then all variables
of the train dataset are used.}
}
\value{
A RasterLayer or data.frame containing the minimum distance to training data.
}
\description{
this function estimates uncertainty of spatial prediction models by
considering the distance of new data (i.e. a Raster Stack of spatial predictors
used in the models) in the predictor variable space to the data used for model
training. Predictors can be weighted in the ideal case based on the internal
variable importance of the machine learning algorithm used for model training.
}
\examples{
library(sf)
library(raster)
library(caret)

# prepare data:
dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
studyArea <- stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))
variables <- c("DEM","NDRE.M","Easting","Northing")
trainDat <- aggregate(dat[,c("VW",variables)],by=list(as.character(dat$SOURCEID)),mean)
studyArea <- studyArea[[which(names(studyArea)\%in\%c(variables))]]

# first calculate uncertainty based on a set of variables with equal weights:
plot(uncertainty(trainDat,studyArea,variables=variables))

# or weight variables based on variable improtance from a trained model:
set.seed(100)
model <- train(trainDat[,which(names(trainDat)\%in\%variables)],
trainDat$VW,method="rf",importance=TRUE)
plot(varImp(model)) # note that coordinates are the major predictors here
plot(uncertainty(trainDat,studyArea,model=model,variables=variables))


}
\author{
Hanna Meyer
}
